backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pydantic==2.6.1
pydantic-settings==2.2.1
pymysql==1.1.1
python-multipart==0.0.9
python-docx==1.1.0
pdfplumber==0.11.0
PyMuPDF==1.24.9
Pillow==11.3.0
pytest==8.2.2
httpx==0.27.0
